added unit tests for new controllers, authentication, authorization, mappers and services

+++ src/main/java/es/thalesalv/chatrpg/infrastructure/inbound/api/response/SearchWorldsResponse.java
@@ -24,6 +24,6 @@ public class SearchWorldsResponse {
     private int page;
     private int totalPages;
     private int resultsInPage;
-    private int totalResults;
+    private long totalResults;
     private List<WorldResponse> results;
 }
